refactor: make datasource not to reference domain layer model
