+++ app/src/main/java/ksc/campus/tech/kakao/map/data/datasources/KakaoSearchService.kt
@@ -12,8 +12,8 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import ksc.campus.tech.kakao.map.BuildConfig
+import ksc.campus.tech.kakao.map.data.entities.Document
 import ksc.campus.tech.kakao.map.data.entities.KeywordSearchResponse
-import ksc.campus.tech.kakao.map.domain.models.SearchResult
 import retrofit2.Call
 import retrofit2.HttpException
 import retrofit2.Response
@@ -39,29 +39,7 @@ class KakaoSearchService @Inject constructor(
 
     private fun isQueryValid(query: String): Boolean = query.isNotBlank()
 
-    private fun parseCategory(category: String) =
-        category.split('>').last().trim().replace(",", ", ")
-
-    private fun responseToResultArray(response: Response<KeywordSearchResponse>): List<SearchResult> {
-        val result = mutableListOf<SearchResult>()
-
-        for (doc in response.body()?.documents ?: listOf()) {
-            result.add(
-                SearchResult(
-                    doc.id,
-                    doc.placeName,
-                    doc.addressName,
-                    CATEGORY_CODE_DESCRIPTION_MAP.getOrDefault(
-                        doc.categoryGroupCode,
-                        parseCategory(doc.categoryName)
-                    ),
-                    doc.y.toDoubleOrNull() ?: 0.0,
-                    doc.x.toDoubleOrNull() ?: 0.0
-                )
-            )
-        }
-        return result
-    }
+    private fun responseToArray(response: Response<KeywordSearchResponse>): List<Document> = response.body()?.documents?: listOf()
 
     private fun isResponseSuccess(response: Response<KeywordSearchResponse>): Boolean {
         if (!response.isSuccessful) {
@@ -76,7 +54,7 @@ class KakaoSearchService @Inject constructor(
     fun searchResult(
         query: String,
         apiKey: String,
-        batchCount: Int): Flow<List<SearchResult>>{
+        batchCount: Int): Flow<List<Document>>{
         return flow {
             try {
                 val result = if (query == "") listOf() else batchSearchByKeyword(
@@ -102,14 +80,14 @@ class KakaoSearchService @Inject constructor(
         apiKey: String,
         page: Int,
         batchCount: Int
-    ): List<SearchResult> {
+    ): List<Document> {
         if (page > batchCount)
             return listOf()
 
         if (!isQueryValid(query))
             return listOf()
 
-        val result = mutableListOf<SearchResult>()
+        val result = mutableListOf<Document>()
         val response =
             retrofitService.requestSearchResultByKeyword("KakaoAK $apiKey", query, page).execute()
 
@@ -117,7 +95,7 @@ class KakaoSearchService @Inject constructor(
             Log.e("KSC", "Message: ${response.message()}")
             return listOf()
         }
-        result += responseToResultArray(response)
+        result += responseToArray(response)
         if (response.body()?.meta?.isEnd == false) {
             result += batchSearchByKeyword(
                 query,
@@ -131,26 +109,6 @@ class KakaoSearchService @Inject constructor(
     }
 
     companion object {
-        private val CATEGORY_CODE_DESCRIPTION_MAP: HashMap<String, String> = hashMapOf(
-            Pair("MT1", "대형마트"),
-            Pair("CS2", "편의점"),
-            Pair("PS3", "어린이집, 유치원"),
-            Pair("SC4", "학교"),
-            Pair("AC5", "학원"),
-            Pair("PK6", "주차장"),
-            Pair("OL7", "주유소, 충전소"),
-            Pair("SW8", "지하철역"),
-            Pair("BK9", "은행"),
-            Pair("CT1", "문화시설"),
-            Pair("AG2", "중개업소"),
-            Pair("PO3", "공공기관"),
-            Pair("AT4", "관광명소"),
-            Pair("AD5", "숙박"),
-            Pair("FD6", "음식점"),
-            Pair("CE7", "카페"),
-            Pair("HP8", "병원"),
-            Pair("PM9", "약국")
-        )
     }
 }
