+++ app/src/main/java/ksc/campus/tech/kakao/map/data/repositoryimpls/SearchResultRepositoryImpl.kt
@@ -1,7 +1,9 @@
 package ksc.campus.tech.kakao.map.data.repositoryimpls
 
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
 import ksc.campus.tech.kakao.map.data.datasources.KakaoSearchService
+import ksc.campus.tech.kakao.map.data.mapper.SearchResponseMapper
 import ksc.campus.tech.kakao.map.domain.models.SearchResult
 import ksc.campus.tech.kakao.map.domain.repositories.SearchResultRepository
 import javax.inject.Inject
@@ -12,7 +14,11 @@ class SearchResultRepositoryImpl @Inject constructor(
     private var text:String = ""
     private var key:String = ""
     override val searchResult: Flow<List<SearchResult>>
-        get() = kakaoSearchService.searchResult(text, key, BATCH_COUNT)
+        get() = kakaoSearchService.searchResult(text, key, BATCH_COUNT).map { document ->
+            document.map {
+                SearchResponseMapper.mapSearchResponseToSearchResult(it)
+            }
+        }
 
     private fun clearResults() {
         text = ""
